Separate firestore-process-delivery code a bit, add logging
